requirements.txt
bcrypt==4.0.1
boto3==1.28.79
botocore==1.31.79
cffi==1.16.0
cryptography==41.0.5
greenlet==3.0.1
jmespath==1.0.1
numpy==1.26.1
pandas==2.1.2
paramiko==3.3.1
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.23
sshtunnel==0.4.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7